- moved testing to jest - adding prettier
