+++ .prettierrc
@@ -1,0 +1,10 @@
+{
+    "printWidth": 110,
+    "tabWidth": 2,
+    "useTabs": false,
+    "semi": true,
+    "singleQuote": false,
+    "bracketSpacing": true,
+    "no-with": 0
+  }
+
