+++ .travis.yml
@@ -1,13 +1,13 @@
 language: node_js
 
 node_js:
-  - "7"
+  - "10"
 
 before-script:
-- npm install
+  - npm install
 
 script:
-- npm run test
+  - npm run test
 
 notifications:
   email: false
