+++ wallaby.conf.js
@@ -1,13 +1,25 @@
 module.exports = function(wallaby) {
   return {
-    files: ["./*.js", "!test/*.spec.js"],
-    tests: ["test/**/*.spec.js"],
+    files: [
+      {
+        pattern: "node_modules/babel-polyfill/dist/polyfill.js",
+        instrument: false
+      },
+      "src/**/*.js",
+      "!test/*.spec.js",
+      "!node_modules/**/*.*"
+    ],
+    tests: ["test/*.spec.js"],
     env: {
       type: "node"
     },
-    testFramework: "mocha",
+    testFramework: "jest",
+    compilers: {
+      "src/**/*.js": wallaby.compilers.babel()
+    },
     workers: {
       recycle: true
-    }
+    },
+    debug: true
   };
 };
